fix(reframe): Phase 3 — pixel-precise subject_x replaces 0-100 int (Bug 4)
Convert ReframeSegment.subject_x from int (0-100 bucket) to float (source
pixel coordinate). This eliminates the ~19.2px quantization error on 1920px
sources. The resolved crop center is now within 1px of the face center.

Changes:
- ReframeSegment: subject_x/subject_y changed from int to float (pixel space)
- _slot_to_x: returns pixel-precise float from face registry's 0-100 x_center
- render_plan_builder._compute_crop_rect: auto-detects pixel vs legacy 0-100
  scale and normalizes to 0.0-1.0 for the Rect output
- pipeline.py: converts pixel-space back to 0-100 int at SceneDescription boundary
- subject_confidence.face_in_proposed_crop: handles pixel-space subject_x
- models_api.py: subject_x changed from int to float
- All test assertions updated for pixel-space values

Pixel rounding happens exactly once: at the FFmpeg export boundary via
Rect.to_pixels() which uses banker's rounding. No rounding anywhere else.

Phase 1 pixel precision tests now pass. All 17 tests green.

+++ backend/models_api.py
@@ -277,7 +277,7 @@ class SceneDeleteResponse(BaseModel):
 class RecenterResponse(BaseModel):
     job_id: str
     scenes_recentered: int
-    subject_x: int = 50
+    subject_x: float = 50.0
     per_scene: bool = False
     message: str = ""
